refactor(app): split transfer consolidation service
